Solve a race condition when used more than once
Solve a race condition caused by MM engine, when the module is used more than once on the page (to show, for example, weather in two different locations).
In such case, `_this` could be of the node_helper of the other module, thus the `url` comes out in the notification not matching the payload, causing the wrong payload to be displayed for the wrong module, and/or the other module stuck on "Loading".

This is caused by the request being caught by the "wrong" node_helper.  Took me considerable debugging to find this one...

So, instead, i changed it to send the actual url stored in the request, attached to the response being analyzed.